Unify Stripe customer.created/updated sync with race-safe create-or-get.
